moves shortcut handling part into separate component
